backend/requirements.txt
manim
langchain
fastapi
pydantic
python-dotenv
langchain[google-genai]
langchain[openai]
langchain[anthropic]
docker
boto3